fix: 테스트 에러 수정

+++ src/test/kotlin/com/beanbliss/domain/user/service/GetBalanceServiceTest.kt
@@ -44,7 +44,7 @@ class GetBalanceServiceTest {
         val mockBalance = BalanceEntity(
             id = 1L,
             userId = userId,
-            amount = 50000,
+            amount = 50000.toBigDecimal(),
             createdAt = now.minusDays(10),
             updatedAt = now
         )
